+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Added:
 - "sites&labs" page has a map which shows listed sites and a button to show a specific listed site on the map
 - possibility to automatically render a created image attachment on the Basic data page
 - devices that are involved in the measurement (loggers, multiplexers) can now be added to the datastream link
+- Include provenance term and ucum case sensitive symbol for cv units
 
 Fixed:
 - loaddata command is now more robust according to files with just a new line